mostly working hehe
