+++ app/assets/stylesheets/desktop/components/_index.scss
@@ -2,4 +2,3 @@
 @import "sidebar/edit-navigation-menu/tags-modal";
 @import "user-card";
 @import "user-info";
-@import "user-stream-item";
