--- app/assets/stylesheets/desktop/components/user-stream-item.scss
@@ -1,8 +1,0 @@
-// Desktop styles for "user-stream-item" component
-.user-stream {
-  .notification {
-    &.unread {
-      background-color: var(--tertiary-low);
-    }
-  }
-}
